Don't show "Camera not authorized" while awaiting permission check
The permission checks in `componentWillMount()` are asynchronous (via Promises), which means that the component can (and usually is) loaded before `this.setState({ isAuthorized })` is able to run.

Thus, `isAuthorized` state will have its original value of `false` at the time of the first `render()` call, and will flash the text `Camera not authorized` while the camera permission check (whether on iOS or Android) is taking place.

This change introduces another state variable `isAuthorizationChecked`, which allows the component to differentiate between camera access being denied vs. awaiting the authorization result, not showing the denial message until the check is fully complete.